Re-organized two functions
