+++ pyutils/data_viz/utils.py
@@ -8,8 +8,8 @@ from pyutils.core.type_annotations import (
     Number,
     NumberOrString,
 )
+from pyutils.core.random_data_generator import generate_random_hex_codes
 from pyutils.core.utils import (
-    generate_random_hex_codes,
     get_max_of_abs_values,
     has_negative_number,
     has_positive_number,
